fix: bound the artifact lock wait by the capture's own budget
Tracing must never spend the budget of the action it observes, and a
trace capture handed the action's own deadline spent all of it waiting
for a contended artifact lock. What reached dispatch was then only the
100 ms the pre-capture path reserves, which quick hardware absorbs and a
loaded shared runner does not.

The wait is now bounded by a fixed 50 ms capture window, clamped to
whatever the caller has left, and a lock that cannot be taken inside it
skips the capture with the reason the trace already carries. The refmap
copy had the same shape with no caller deadline at all, and takes the
same window through the ambient one.

The contended test stops proving this with a stopwatch: it reads the
budget the action actually dispatched with, that the capture was skipped
and traced as such, and that dispatch happened exactly once.
